+++ CHANGES.md
@@ -1,5 +1,10 @@
 # **libCLImate.Ruby** Changes
 
+## 0.14.0 - 15th April 2019
+
+* + Climate#constrain_values now supports Array type (as well as Integer and Range)
+* + added Climate#usage_help_suffix, which defaults to "use --help for usage"
+
 ## 0.13.0 - 13th April 2019
 
 * + added CLASP::Arguments.load(), which allows to load argument-specifications from Hash or from YAML